requirements.txt
async-generator==1.10
attrs==22.1.0
behave==1.2.6
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==22.0
parse==1.19.0
parse-type==0.6.0
PyHamcrest==2.0.4
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.7.2
selenium-page-factory==2.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0
